+++ api/app/api/router.py
@@ -1,5 +1,5 @@
 from fastapi import APIRouter
-from api.routes import ia, auth, conversations, admin
+from api.routes import ia, auth, conversations, admin, mfa
 
 api_router = APIRouter()
 
@@ -7,3 +7,4 @@ api_router.include_router(auth.router)
 api_router.include_router(ia.router)
 api_router.include_router(conversations.router)
 api_router.include_router(admin.router)
+api_router.include_router(mfa.router)